==> notebooks/00.CleanText.ipynb <==
import os
import re
import csv
import pandas as pd

from tqdm import tqdm

!ls /eee/tgnews/meta/all

Папка */eee/tgnews/meta/all* содержит данные извлечённые из оригинальных *.xml.

Далее отчистим тексты и запишим их в папку */eee/tgnews/meta/all_clean*.

def clean(text):
    
    text = re.sub('https*:\S*', ' ', text) #удаляем ссылки
    text = re.sub('[^\w\.\!\?\(\)\-:;,]+', ' ', text) #удаляем все символы, кроме букв, цифр и знаков пунктуации
    text = re.sub('\s+', ' ', text) #заменяем все последовательности пробельных символов на один пробел
    
    return text.strip()

load_path = "/eee/tgnews/meta/all"
dump_path = "/eee/tgnews/meta/all_clean"

columns = ["og:title", "og:description", "text"]

for table_name in tqdm(os.listdir(load_path)):
    
    table = pd.read_csv(os.path.join(load_path, table_name), sep='\t', keep_default_na=False, quoting=csv.QUOTE_NONE)
    
    for column in columns:
        table[column] = table[column].apply(clean)
    
    table.to_csv(os.path.join(dump_path, table_name), sep='\t', index=False, quoting=csv.QUOTE_NONE)

table.head(1)
