==> notebooks/01.DetectLangs.ipynb <==
import os
import csv
import cld3
import pandas as pd

from tqdm import tqdm
from itertools import chain

tqdm.pandas()

!ls /eee/tgnews/meta/all_clean

def detect_lang(text):
    
    lang_pred = cld3.get_language(text)
    
    if lang_pred is not None:
        return lang_pred.language
    
    return None

load_path = "/eee/tgnews/meta/all_clean"
dump_path = "/eee/tgnews/meta/all_lang"

for table_name in os.listdir(load_path):
    
    table = pd.read_csv(os.path.join(load_path, table_name), sep='\t', keep_default_na=False, quoting=csv.QUOTE_NONE)
    table["lang"] = table["text"].progress_apply(detect_lang)
    
    table.to_csv(os.path.join(dump_path, table_name), sep='\t', index=False, quoting=csv.QUOTE_NONE)

table.head(1)

#Считаем количество всевозможных языков для каждого источника

load_path = "/eee/tgnews/meta/all_lang"

site_langs = {}
for table_name in tqdm(os.listdir(load_path)):
    
    table = pd.read_csv(os.path.join(load_path, table_name), usecols=["path", "og:site_name", "lang"], sep='\t', quoting=csv.QUOTE_NONE)
    table = table.dropna()
    
    for (site_name, lang), count in table.groupby(["og:site_name", "lang"])["path"].count().iteritems():
        
        site_langs.setdefault(site_name, {})
        
        if lang not in {"ru", "en"}:
            lang = "other"
        
        site_langs[site_name].setdefault(lang, 0)
        site_langs[site_name][lang] += count

#Оставляем только те источники, которые содержат более статей

site_langs = {site_name:langs for site_name, langs in site_langs.items() if sum(langs.values()) >= 250}
print(f"Всего {len(site_langs)} таких источников")

#Нормируем число статей каждого языка для каждого источника

for site_name, langs in tqdm(site_langs.items()):
    
    sum_count = sum(langs.values())
    site_langs[site_name] = {lang:count/sum_count for lang, count in langs.items()}

#Нормируем списки источников, для которых мы уверены в языке

sites = {"ru":[], "en":[], "other":[]}

for site_name, langs in site_langs.items():
    for lang, weight in langs.items():
        
        if weight > 0.99:
            sites[lang].append(site_name)
            break

#Сохраняем списки

for lang, lang_sites in sites.items():
    with open(f"../lang_list/{lang}.txt", "w") as fl:
        fl.write('\n'.join(lang_sites))

!realpath ../lang_list

sites = dict(chain(*[[(site, lang) for site in sites] for lang, sites in sites.items()]))

#Скоректируем языки с учётом списков

load_path = "/eee/tgnews/meta/all_lang"
dump_path = "/eee/tgnews/meta/all_lang_list"

for table_name in tqdm(os.listdir(load_path)):
    
    table = pd.read_csv(os.path.join(load_path, table_name), sep='\t', keep_default_na=False, quoting=csv.QUOTE_NONE)
    
    table["lang"] = table["lang"].apply(lambda x: x if x in {"ru", "en"} else "other")
    table["lang"] = table.apply(lambda x: sites[x["og:site_name"]] if x["og:site_name"] in sites else x["lang"] , axis=1)
    
    table.to_csv(os.path.join(dump_path, table_name), sep='\t', index=False, quoting=csv.QUOTE_NONE)

#Разделяем на языки и сохраняем в разные таблицы

load_path = "/eee/tgnews/meta/all_lang_list"
dump_path = "/eee/tgnews/meta/"

table_ru = []
table_en = []

for table_name in tqdm(os.listdir(load_path)):
    
    table = pd.read_csv(os.path.join(load_path, table_name), sep='\t', keep_default_na=False, quoting=csv.QUOTE_NONE)
    
    table_ru.append(table[table["lang"] == "ru"])
    table_en.append(table[table["lang"] == "en"])
    
table_ru = pd.concat(table_ru, axis=0, ignore_index=True, sort=False).drop(columns="lang")
table_en = pd.concat(table_en, axis=0, ignore_index=True, sort=False).drop(columns="lang")

table_ru.to_csv(os.path.join(dump_path, "ru.tsv"), sep='\t', index=False, quoting=csv.QUOTE_NONE)
table_en.to_csv(os.path.join(dump_path, "en.tsv"), sep='\t', index=False, quoting=csv.QUOTE_NONE)

len(table_ru), len(table_en)
