==> notebooks/03.GetClasses.ipynb <==
import re
import os
import csv
import artm
import numpy as np
import pandas as pd

from tqdm import tqdm
from itertools import chain 
from collections import Counter
from ufal.udpipe import Model, Sentence, ProcessingError

load_path = "/eee/tgnews/meta/"

columns = ["path", "og:site_name", "og:url"]

table_ru = pd.read_csv(os.path.join(load_path, "ru.tsv"), sep='\t', usecols=columns, keep_default_na=False, quoting=csv.QUOTE_NONE)
table_en = pd.read_csv(os.path.join(load_path, "en.tsv"), sep='\t', usecols=columns, keep_default_na=False, quoting=csv.QUOTE_NONE)

len(table_ru), len(table_en)

#Выделение токенов из url

def url_to_tokens(url, lowercase=True, split_pattern="[^a-z]+", filter_pattern="[a-z]{3,}"):
    
    tokens = re.sub("https*://[^/]*", "", url).strip("/").split("/")[:-1]
    
    if lowercase:
        tokens = [token.lower() for token in tokens]
        
    if split_pattern is not None:
        tokens = list(chain(*[re.split(split_pattern, token) for token in tokens]))
        
    if filter_pattern is not None:
        tokens = list(filter(lambda x: re.search(filter_pattern, x), tokens))
    
    return tokens

table_ru["tokens:url"] = table_ru["og:url"].apply(url_to_tokens)
table_en["tokens:url"] = table_en["og:url"].apply(url_to_tokens)

#Формируем список наиболее частотных токенов

site_tokens = dict([(site, Counter(tokens)) for site, tokens in table_ru.groupby("og:site_name")["tokens:url"].sum().iteritems()])

for site, tokens in table_en.groupby("og:site_name")["tokens:url"].sum().iteritems():
    
    site_tokens.setdefault(site, Counter())
    site_tokens[site] += Counter(tokens)
    
token_count = {}
token_sites = {}

for site, tokens in site_tokens.items():
    for token, count in tokens.items():
        
        token_count.setdefault(token, 0)
        token_count[token] += count
        
        token_sites.setdefault(token, [])
        token_sites[token].append(site)
        
tokens = [token for token, count in token_count.items() if (count >= 100) and (len(token_sites[token]) >= 10)]

with open("../url_tokens.txt", "w") as fl:
    fl.write('\n'.join(tokens))

def predict_class_news_by_url(url_tokens):
    
    url_tokens = " ".join(url_tokens)
    classes = []
    
    for part in ["society",
                 "politic",
                 "incident",
                 "criminal",
                 "proisshestviya",
                 "obshchestvo",
                 "social",
                 "murder",
                 "politika",
                 "crime",
                 "kriminal",
                 "parliament",
                 "president",
                 "law",
                 "government",
                 "political"]:
    
        if part in url_tokens:
            classes.append("Society")
            break
        
    for part in ["econom",
                 "ekonomika",
                 "business",
                 "industry",
                 "biznes",
                 "bank",
                 "energy",
                 "company",
                 "finance"]:
    
        if part in url_tokens:
            classes.append("Economy")
            break
        
    for part in ["sport",
                 "hockey",
                 "football",
                 "tennis",
                 "basketball",
                 "box",
                 "fitness",
                 "athletic",
                 "golf"]:
        
        if part in url_tokens:
            classes.append("Sports")
            break
        
    for part in ["auto",
                 "tehn",
                 "internet",
                 "digital",
                 "mobile",
                 "gadgets",
                 "smart"]:
        
        if part in url_tokens:
            classes.append("Technology")
            break
        
    for part in ["art",
                 "kino",
                 "culture",
                 "kultura",
                 "entertainment",
                 "afisha",
                 "cinema",
                 "book",
                 "game",
                 "music",
                 "party",
                 "gaming"]:
        
        if part in url_tokens:
            classes.append("Entertainment")
            break
    
    for part in ["nauka",
                 "health",
                 "research",
                 "science",
                 "medicine"]:
        
        if part in url_tokens:
            classes.append("Science")
            break
    
    if not classes:
        return None
    
    return classes

table_ru["class"] = table_ru["tokens:url"].apply(predict_class_news_by_url)
table_en["class"] = table_en["tokens:url"].apply(predict_class_news_by_url)

url_token = "features"

paths = table_ru[table_ru["tokens:url"].apply(lambda x: url_token in x)]["path"].tolist()
paths = paths + table_en[table_en["tokens:url"].apply(lambda x: url_token in x)]["path"].tolist()

with open(paths[np.random.choice(range(len(paths)))]) as fl:
    print(fl.read())

Counter(chain(*table_ru["class"].dropna())).most_common()

Counter(chain(*table_en["class"].dropna())).most_common()

**ru**

table_ru = table_ru.join(pd.read_csv(os.path.join(load_path, "ru_preproc.csv"), keep_default_na=False).set_index("path"), on="path")

udpipe_model = Model.load('/eee/tgnews/misc/ru.udpipe')
conllu_tokenizer = udpipe_model.newTokenizer("conllu").newConlluInputFormat()

def conllu_encode(conllu_text):
    
    sentence = Sentence()
    error = ProcessingError()
    
    conllu_tokenizer.setText(conllu_text)
    
    sentences = []
    while conllu_tokenizer.nextSentence(sentence, error):
        sentences.append(sentence.words[1:]) 
        
    return sentences

table_ru["og:title"] = table_ru["og:title"].apply(conllu_encode)
table_ru["text"] = table_ru["text"].apply(conllu_encode)

def vowpal_wabbit_encode(sents):
    
    upostags = {'ADJ', 'INTJ', 'NOUN', 'PROPN', 'VERB'}
    
    tokens = [[token.lemma for token in sent if token.upostag in upostags] for sent in sents]
    tokens = [token.replace(":", "") for token in chain(*tokens)]
    tokens = Counter(tokens)
    tokens = [token + ('' if count == 1 else f':{count}') for token, count in tokens.items()]
    
    return ' '.join(tokens)

table_ru["og:title"] = table_ru["og:title"].apply(vowpal_wabbit_encode)
table_ru["text"] = table_ru["text"].apply(vowpal_wabbit_encode)

with open("../table_ru_vw.txt", "w") as fl:
    fl.write('\n'.join(table_ru.apply(lambda x: x["path"] + " |title_tokens " + x["og:title"] + " |text_tokens " + x["text"], axis=1)) + '\n')
    
batch_vectorizer = artm.BatchVectorizer(data_path='../table_ru_vw.txt', 
                                        data_format='vowpal_wabbit',
                                        target_folder='../batches_ru')

batch_vectorizer.dictionary.filter(min_df=3, min_tf=5)

batch_vectorizer.dictionary.save_text(dictionary_path='../models/dictionary_ru.txt')
batch_vectorizer.dictionary.save(dictionary_path='../models/dictionary_ru.dict')

num_topics = 8
num_tokens = 10

class_ids = ["title_tokens", "text_tokens"]
topic_names=["Society", "Economy", "Sports", "Technology", "Entertainment", "Science", "Other", "background"]

docs_for_sst = table_ru[~table_ru["class"].isna()][["path", "class"]]
n_docs = len(docs_for_sst)

doc_titles = docs_for_sst.path.tolist()
topic_indecies = docs_for_sst["class"].apply(lambda x: [topic_names.index(topic) for topic in x]).tolist()

doc_topic_coef = np.zeros((n_docs, num_topics))
for index, topic_index in enumerate(topic_indecies):
    doc_topic_coef[index, topic_index] = 1
    
doc_topic_coef = doc_topic_coef / doc_topic_coef.sum(axis=1)[:, np.newaxis]
doc_topic_coef = doc_topic_coef.tolist()

model = artm.ARTM(num_topics=num_topics, 
                  class_ids=class_ids,
                  topic_names=topic_names,
                  dictionary=batch_vectorizer.dictionary, 
                  theta_columns_naming="title")

model.regularizers.add(artm.SmoothSparseThetaRegularizer(name='SST', tau=100, 
                                                         doc_titles=doc_titles, 
                                                         doc_topic_coef=doc_topic_coef))

model.regularizers.add(artm.SmoothSparsePhiRegularizer('SmoothPhi_back', tau=1, gamma=0, 
                                                       topic_names=["background"]))

model.regularizers.add(artm.DecorrelatorPhiRegularizer('DecPhi', 
                                                       tau=0.1, 
                                                       gamma=0))

for class_id in class_ids:
    model.scores.add(artm.TopTokensScore(name=f'TT_{class_id}', class_id=class_id, num_tokens=num_tokens))
    
model.fit_offline(batch_vectorizer=batch_vectorizer, num_collection_passes=5)

for modality, modality_top_tokens_tracker in model.score_tracker.items():
    print(modality)
    for topic_name, modality_top_tokens in modality_top_tokens_tracker.last_tokens.items():
        print(topic_name)
        print('; '.join(modality_top_tokens))
    print()

model.dump_artm_model('../models/cl_news_model_ru.dump_tm')
model.save('../models/cl_news_model_ru.save_tm')

**en**

table_en = table_en.join(pd.read_csv(os.path.join(load_path, "en_preproc.csv"), keep_default_na=False).set_index("path"),  on="path")

table_en["og:title"] = table_en["og:title"].apply(conllu_encode)
table_en["text"] = table_en["text"].apply(conllu_encode)

table_en["og:title"] = table_en["og:title"].apply(vowpal_wabbit_encode)
table_en["text"] = table_en["text"].apply(vowpal_wabbit_encode)

with open("../table_en_vw.txt", "w") as fl:
    fl.write('\n'.join(table_en.apply(lambda x: x["path"] + " |title_tokens " + x["og:title"] + " |text_tokens " + x["text"], axis=1)) + '\n')
    
batch_vectorizer = artm.BatchVectorizer(data_path='../table_en_vw.txt', 
                                        data_format='vowpal_wabbit',
                                        target_folder='../batches_en')

batch_vectorizer.dictionary.filter(min_df=3, min_tf=5)

batch_vectorizer.dictionary.save_text(dictionary_path='../models/dictionary_en.txt')
batch_vectorizer.dictionary.save(dictionary_path='../models/dictionary_en.dict')

num_topics = 8
num_tokens = 10

class_ids = ["title_tokens", "text_tokens"]
topic_names=["Society", "Economy", "Sports", "Technology", "Entertainment", "Science", "Other", "background"]

docs_for_sst = table_en[~table_en["class"].isna()][["path", "class"]]
n_docs = len(docs_for_sst)

doc_titles = docs_for_sst.path.tolist()
topic_indecies = docs_for_sst["class"].apply(lambda x: [topic_names.index(topic) for topic in x]).tolist()

doc_topic_coef = np.zeros((n_docs, num_topics))
for index, topic_index in enumerate(topic_indecies):
    doc_topic_coef[index, topic_index] = 1
    
doc_topic_coef = doc_topic_coef / doc_topic_coef.sum(axis=1)[:, np.newaxis]
doc_topic_coef = doc_topic_coef.tolist()

model = artm.ARTM(num_topics=num_topics, 
                  class_ids=class_ids,
                  topic_names=topic_names,
                  dictionary=batch_vectorizer.dictionary, 
                  theta_columns_naming="title")

model.regularizers.add(artm.SmoothSparseThetaRegularizer(name='SST', tau=100, 
                                                         doc_titles=doc_titles, 
                                                         doc_topic_coef=doc_topic_coef))

model.regularizers.add(artm.SmoothSparsePhiRegularizer('SmoothPhi_back', tau=1, gamma=0, 
                                                       topic_names=["background"]))

model.regularizers.add(artm.DecorrelatorPhiRegularizer('DecPhi', 
                                                       tau=0.1, 
                                                       gamma=0))

for class_id in class_ids:
    model.scores.add(artm.TopTokensScore(name=f'TT_{class_id}', class_id=class_id, num_tokens=num_tokens))
    
model.fit_offline(batch_vectorizer=batch_vectorizer, num_collection_passes=5)

for modality, modality_top_tokens_tracker in model.score_tracker.items():
    print(modality)
    for topic_name, modality_top_tokens in modality_top_tokens_tracker.last_tokens.items():
        print(topic_name)
        print('; '.join(modality_top_tokens))
    print()

model.dump_artm_model('../models/cl_news_model_en.dump_tm')
model.save('../models/cl_news_model_en.save_tm')
