==> notebooks/04.SelectEvents.ipynb <==
import artm

from collections import Counter

batch_vectorizer = artm.BatchVectorizer(data_path='../batches_ru', data_format='batches')

dictionary = artm.Dictionary()
dictionary.load(dictionary_path='../models/dictionary_ru.dict')

model = artm.ARTM(num_topics=1000, 
                  class_ids=["title_tokens", "text_tokens"], 
                  dictionary=dictionary, 
                  theta_columns_naming="title", 
                  cache_theta=True)

model.regularizers.add(artm.TopicSelectionThetaRegularizer(tau=0.3, topic_names=model.topic_names[:-1]))

model.regularizers.add(artm.SmoothSparsePhiRegularizer('SmoothPhi_back', tau=1, gamma=0, topic_names=model.topic_names[-1]))
model.regularizers.add(artm.DecorrelatorPhiRegularizer('DecPhi', tau=0.1, gamma=0))

model.fit_offline(batch_vectorizer=batch_vectorizer, num_collection_passes=5)

predict = model.get_theta().T.idxmax(axis=1)

Counter(predict.values)
